refactor: 결과 계산 클래스 분리
