+++ src/step1/app.js
@@ -3,8 +3,8 @@ import Money from "./model/Money.js";
 import LottoStore from "./model/LottoStore.js";
 import WinningLotto from "./model/WinningLotto.js";
 import { INPUT_MESSAGE, ERROR_MESSAGE } from "./constant/message.js";
-import { parseNumbers, evaluateLotto } from "./utils.js";
-import { getCountsRank, getReturnOnInvestment, printResult } from "./utils.js";
+import { parseNumbers } from "./utils.js";
+import { getReturnOnInvestment, printResult } from "./utils.js";
 import Lotto from "./model/Lotto.js";
 
 class App {
@@ -18,12 +18,11 @@ class App {
 
   async run() {
     const money = await this.#askMoney();
-    const lottoCount = money.getLottoCount(); // money / LOTTO.PRICE
+    const lottoCount = money.getLottoCount();
 
-    // const lottoStore = new LottoStore();
     const lottos = this.#lottoStore.issuedLotto(lottoCount);
 
-    // 발행한 로또 출력a
+    // Output
     console.log(`${money.getLottoCount()}장을 구매했습니다.`);
     lottos.forEach((lotto) =>
       console.log(`[${lotto.getNumbers().join(", ")}]`),
@@ -35,15 +34,14 @@ class App {
     // 당첨 '보너스 번호' 입력받기(번호)
     const winningLotto = await this.#askBonusNumber(winningNumbers);
 
-    // 구입한 로또 등수 계산   { RANK_1: 2, RANK_2: 1, RANK_3: 1, RANK_4: 0, RANK_5: 1 }
-    const ranks = lottos.map((lotto) => evaluateLotto(lotto, winningLotto));
+    // 구입한 로또 등수 계산
+    const result = winningLotto.evaluateLottos(lottos);
 
     // 수익률 계산
-    const counts = getCountsRank(ranks);
-    const returnOnInvestment = getReturnOnInvestment(money.getMoney(), counts);
+    const returnOnInvestment = result.getReturnOnInvestment(money.getMoney());
 
     // 당첨 통계 출력
-    printResult(counts, returnOnInvestment);
+    printResult(result.getCounts(), returnOnInvestment);
 
     // 다시 시작 여부 묻기
     await this.#askRetry();
